feature: allow files with more than 2147483647 lines (#339)
This is WIP. Indexing and searching works, but some places where Qt
requires to use int numbers (scrollbars, spinboxes etc.) are buggy.

+++ src/ui/src/crawlerwidget.cpp
@@ -60,6 +60,7 @@
 #include <QShortcut>
 #include <QStandardItemModel>
 #include <QStringListModel>
+#include <qglobal.h>
 
 #include "data/regularexpression.h"
 
@@ -1529,7 +1530,7 @@ QString CrawlerWidgetContext::toString() const
     const auto toVariantList = []( const auto& list ) -> QVariantList {
         QVariantList variantList;
         for ( const auto& item : list ) {
-            variantList.append( item );
+            variantList.append( static_cast<qulonglong>(item) );
         }
         return variantList;
     };
